OR-532: FraudCheck now takes FraudAssessmentRequest, an object which has the requestVersion and the order object, as a parameter

+++ src/test/groovy/ca/bestbuy/orders/fraud/mappers/TASResponseXMLMapperTest.groovy
@@ -62,7 +62,7 @@ class TASResponseXMLMapperTest extends Specification{
         fraudResult.getOrderNumber() == transactionResults.getTransactionId()
         fraudResult.getTotalFraudScore() == transactionResults.getTotalScore()
         fraudResult.getRecommendationCode() == transactionResults.getRecommendationCode()
-        fraudResult.getRequestVersion() == transactionResults.getResponseData().getResponseVersion()
+        fraudResult.getRequestVersion() == transactionResults.getResponseData().getResponseVersion().toLong()
 
     }
